# docs/swagger.yaml
basePath: /api/v1
definitions:
  ai.ChatMessage:
    properties:
      content:
        type: string
      parts:
        items:
          $ref: '#/definitions/ai.ChatMessagePart'
        type: array
      role:
        type: string
    type: object
  ai.ChatMessagePart:
    properties:
      text:
        type: string
      type:
        type: string
    type: object
  ai.ChatRequest:
    properties:
      conversationId:
        type: integer
      id:
        type: string
      messages:
        items:
          $ref: '#/definitions/ai.ChatMessage'
        type: array
      projectId:
        type: string
    type: object
  ai.ConversationDetail:
    properties:
      id:
        type: integer
      messages:
        items:
          $ref: '#/definitions/ai.Message'
        type: array
      projectId:
        type: integer
      startedAt:
        type: string
    type: object
  ai.ConversationResponse:
    properties:
      id:
        type: integer
      projectId:
        type: integer
      startedAt:
        type: string
    type: object
  ai.CreateConversationRequest:
    properties:
      title:
        type: string
    type: object
  ai.Message:
    properties:
      content:
        type: string
      conversationId:
        type: integer
      createdAt:
        type: string
      id:
        type: integer
      sender:
        type: string
      toolCalls:
        items:
          $ref: '#/definitions/db.ToolCall'
        type: array
    type: object
  ai.Model:
    properties:
      description:
        type: string
      maxTokens:
        type: integer
      name:
        type: string
    type: object
...
  boilerplates.BoilerplateConfig:
    properties:
      args:
        items:
          type: string
        type: array
      command:
        type: string
      description:
        type: string
      id:
        description: ID is the key in the configs map
        type: string
      image:
        type: string
      name:
        type: string
      platform:
        type: string
      repoName:
        type: string
      repoOwner:
        type: string
      repoPath:
        type: string
    type: object
...
  db.ToolCall:
    properties:
      arguments:
        type: string
      createdAt:
        type: string
      error:
        $ref: '#/definitions/sql.NullString'
      id:
        type: integer
      messageId:
        type: integer
      result:
        $ref: '#/definitions/sql.NullString'
      toolName:
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  sql.NullString:
    properties:
      string:
        type: string
      valid:
        description: Valid is true if String is not NULL
        type: boolean
    type: object
...
host: localhost:8100
info:
  contact:
    email: support@chainlaunch.dev
    name: API Support
    url: http://chainlaunch.dev/support
  description: ChainLaunch API provides services for managing blockchain networks
    and cryptographic keys
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ChainLaunch API
  version: "1.0"
paths:
  /ai/{projectId}/chat:
    post:
      consumes:
      - application/json
      description: Stream a conversation with the AI assistant using Server-Sent Events
        (SSE)
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: integer
      - description: Chat request containing project ID and messages
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ai.ChatRequest'
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE stream of chat responses
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Chat with AI assistant
      tags:
      - ai
  /ai/{projectId}/conversations:
    get:
      description: Returns a list of all chat conversations associated with a specific
        project
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ai.ConversationResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get all conversations for a project
      tags:
      - ai
    post:
      consumes:
      - application/json
      description: Creates a new empty conversation for the specified project
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: integer
      - description: Optional conversation title
        in: body
        name: request
        schema:
          $ref: '#/definitions/ai.CreateConversationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ai.ConversationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create a new conversation for a project
      tags:
      - ai
  /ai/{projectId}/conversations/{conversationId}:
    get:
      description: Get all messages in a conversation
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: integer
      - description: Conversation ID
        in: path
        name: conversationId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ai.Message'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get conversation messages
      tags:
      - ai
  /ai/{projectId}/conversations/{conversationId}/export:
    get:
      description: Get detailed information about a conversation including all messages
        and metadata
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: integer
      - description: Conversation ID
        in: path
        name: conversationId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ai.ConversationDetail'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get conversation detail
      tags:
      - ai
  /ai/boilerplates:
    get:
      description: Returns a list of available boilerplates filtered by network platform
      parameters:
      - description: Network ID to filter boilerplates by platform
        in: query
        name: network_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/boilerplates.BoilerplateConfig'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get available boilerplates
      tags:
      - ai
  /ai/models:
    get:
      description: Returns a list of available AI models for code generation
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ai.Model'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get available AI models
      tags:
      - ai
...
schemes:
- http
- https
securityDefinitions:
  BasicAuth:
    type: basic
  CookieAuth:
    in: cookie
    name: session_id
    type: apiKey
swagger: "2.0"
...